+++ drivers/gpu/drm/msm/msm_gem.c
@@ -338,7 +338,7 @@ static struct msm_gem_vma *obj_get_domain(struct drm_gem_object *obj,
  * the refcnt counter needs to be atomic_t.
  */
 int msm_gem_get_iova_locked(struct drm_gem_object *obj,
-		struct msm_gem_address_space *aspace, uint32_t *iova)
+		struct msm_gem_address_space *aspace, uint64_t *iova)
 {
 	struct msm_gem_object *msm_obj = to_msm_bo(obj);
 	struct page **pages;
@@ -351,7 +351,7 @@ int msm_gem_get_iova_locked(struct drm_gem_object *obj,
 		if (IS_ERR(pages))
 			return PTR_ERR(pages);
 
-		*iova = physaddr(obj);
+		*iova = (uint64_t) physaddr(obj);
 		return 0;
 	}
 
@@ -382,7 +382,7 @@ int msm_gem_get_iova_locked(struct drm_gem_object *obj,
 
 /* get iova, taking a reference.  Should have a matching put */
 int msm_gem_get_iova(struct drm_gem_object *obj,
-		struct msm_gem_address_space *aspace, uint32_t *iova)
+		struct msm_gem_address_space *aspace, uint64_t *iova)
 {
 	struct msm_gem_vma *domain;
 	int ret;
@@ -402,7 +402,7 @@ int msm_gem_get_iova(struct drm_gem_object *obj,
 /* get iova without taking a reference, used in places where you have
  * already done a 'msm_gem_get_iova()'.
  */
-uint32_t msm_gem_iova(struct drm_gem_object *obj,
+uint64_t msm_gem_iova(struct drm_gem_object *obj,
 		struct msm_gem_address_space *aspace)
 {
 	struct msm_gem_vma *domain = obj_get_domain(obj, aspace);
